refactor: move Toolbar
