+++ src/web/src/views/workspace/WSEditor/WSEditor.tsx
@@ -3,7 +3,7 @@ import { Box, Dialog, Slide, Drawer, Toolbar } from "@mui/material";
 import { useParams } from "react-router";
 import { TransitionProps } from "@mui/material/transitions";
 import WSEditorSwaggerPicker from "../WSEditorSwaggerPicker";
-import WSEditorToolBar from "../WSEditorToolBar";
+import WSEditorToolBar from "./WSEditorToolBar";
 import WSEditorCommandTree, { CommandTreeLeaf, CommandTreeNode } from "./WSEditorCommandTree";
 import WSEditorCommandGroupContent, { DecodeResponseCommandGroup } from "../WSEditorCommandGroupContent";
 import WSEditorCommandContent, { DecodeResponseCommand } from "../WSEditorCommandContent";
