+++ src/web/src/views/workspace/WSEditor/index.ts
@@ -1,4 +1,5 @@
 export { WSEditor, WSEditor as default } from "./WSEditor";
+export { default as WSEditorToolBar } from "./WSEditorToolBar";
 export { default as WSEditorCommandTree } from "./WSEditorCommandTree";
 export type { CommandTreeNode, CommandTreeLeaf } from "./WSEditorCommandTree";
 export { default as WSEditorExportDialog } from "./WSEditorExportDialog";
